proxmox-web-ui/requirements.txt
Flask
proxmoxer
python-dotenv